WIP: pre-test checkpoint — split routes/detectors/store.py
