+++ CLAUDE.md
@@ -97,7 +97,7 @@ If a failure is genuinely outside the scope of the current task (e.g. a flaky ne
 - `vtsearch/settings.py` — Persistent settings, split across two tiers. **Server tier** (shared, stored in `data/settings.json` at `SETTINGS_PATH`): saved_datasets_dir, detectors_dir, max_concurrent_dataset_downloads, max_concurrent_dataset_embeddings, autorun_detectors. **Per-user tier** (stored in `<get_user_data_dir(user)>/user_settings.json`, isolated per user via `get_current_user()`): volume, inclusion, theme, enrich_descriptions, safe_thresholds, calibrate_count, calibration_fraction, audio_playing, swipe_animation, show_metadata, view_mode_left/right, focus_mode_left/right, grid_icon_size_left/right, panel_pct_left/right, autopilot_enabled, hide_autopilot, autopilot_top_greens/hard_reds/resort_interval/goal_diversity, settings_source, achievement_state. Each `set_*` accessor routes to the correct tier; `_save_user()` also auto-syncs to the user's configured `settings_source` (guarded by a per-user entry in `_syncing` to prevent circular re-export). `sync_from_settings_source()` runs lazily on each user's first per-user-cache load each process lifetime (tracked in `_synced_users`). `get_user_settings()` returns only the current user's keys for export. Pre-refactor `data/settings.json` files containing per-user keys are migrated into `data/default/user_settings.json` on first load. Background threads spawned from a request handler must call `vtsearch.auth.set_thread_user()` so per-user writes resolve correctly (done automatically by `JobManager` and the dataset-load/stage thread spawn sites).
 - `vtsearch/routes/` — Flask blueprints, grouped by domain into sub-packages with flat top-level files for single-blueprint domains:
   - `datasets/` — `crud.py` (import/staging/media-types), `registry.py` (registry CRUD at `/api/datasets/registry/*`), `ui.py` (dashboard, demo list, display name)
-  - `detectors/` — `store.py` (on-disk labelset+query store at `/api/detectors/*`), `registry.py` (in-memory detector registry at `/api/detectors/registry/*`, autorun toggle), `scoring.py` (`/api/auto-detect`, `/api/find-label`), `find.py` (multi-dataset Find)
+  - `detectors/` — `crud.py` (detector CRUD: list/create/get/delete/rename/examples/combine at `/api/detectors/*`), `labels.py` (labelset save/import/labels-detail/preview/thumbnail/vote at `/api/detectors/<name>/labels*`), `registry.py` (in-memory detector registry at `/api/detectors/registry/*`, autorun toggle), `scoring.py` (`/api/auto-detect`, `/api/find-label`), `find.py` (multi-dataset Find)
   - `processors/` — `crud.py` and `scoring.py` for extractor/localizer/pregen-processor CRUD + execution
   - `media/` — `list.py` (media-list APIs), `server.py` (server-side media files + example-sort), `embed.py` (on-demand embedding)
   - `settings/` — `api.py` (settings CRUD), `io.py` (import/export), `sources.py` (sync sources)
