+++ docs/ARCHITECTURE.md
@@ -574,8 +574,8 @@ See `_DEFAULTS` in `settings.py` for the full list.
 Theme supports three modes: `dark`, `light`, and `highviz` (high-contrast).
 
 Trainable models are persisted as JSON files in `data/detectors/`
-via the `detectors_bp` route blueprint.  Each stores a name,
-text query, media type, examples list, and labelset.
+via the `detectors_crud_bp` / `detectors_labels_bp` route blueprints.
+Each stores a name, text query, media type, examples list, and labelset.
 
 **Primarily Flask routes mutate this state.**  Most ML and dataset
 functions accept state as parameters — so you can use the ML code in a
